Refactor main.ts to handle non-iterable response.data
Update the main.ts file to handle the case where response.data is not iterable. If response.data is an array, push its contents to seatsData. Otherwise, handle the non-iterable response.data by logging an error message. This change improves the robustness of the code when dealing with different types of response data.